% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatInput_reps.R
\docType{data}
\name{BinMatInput_reps}
\alias{BinMatInput_reps}
\title{Example input data containing a binary matrix comprising replicate pairs}
\format{A dataframe with columns for loci, and rows of replicate pairs.}
\usage{
data(BinMatInput_reps)
}
\description{
Example input data containing a binary matrix comprising replicate pairs
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
check_data(mat)
pks = peaks(mat)
cons = consolidate(mat)
err = errors(cons)
rem = peakRemove(cons, 4)
clust = upgma(cons)


}
\keyword{datasets}
