% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunias_orientalis.R
\docType{data}
\name{bunias_orientalis}
\alias{bunias_orientalis}
\title{Example input file of Bunias orientalis AFLP data, taken from Tewes et. al. (2017). This dataset has already been consolidated, and can be used
as input for the generation of an nMDS plot. The paper can be found here: <doi:10.1111/1365-2745.12869>}
\format{
A dataframe with columns for loci, and rows of replicate pairs. Grouping information is in the second column.
}
\usage{
data(bunias_orientalis)
}
\description{
Example input file of Bunias orientalis AFLP data, taken from Tewes et. al. (2017). This dataset has already been consolidated, and can be used
as input for the generation of an nMDS plot. The paper can be found here: <doi:10.1111/1365-2745.12869>
}
\examples{
bunias = bunias_orientalis
group.names(bunias)
nmds(bunias, labs = FALSE, include_ellipse = TRUE, legend_pos = "right")


}
\keyword{datasets}
