% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check.data}
\alias{check.data}
\title{Checks binary matrix for unwanted characters.}
\usage{
check.data(x)
}
\arguments{
\item{x}{A CSV file containing replicate pairs of binary data.}
}
\value{
Index positions where unwanted values occur (row, column).
}
\description{
Checks for unwanted values (other than 1, 0, and ?).
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
check.data(mat)

}
