% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors}
\alias{errors}
\title{Calculates Jaccard and Euclidean error rates.}
\usage{
errors(x)
}
\arguments{
\item{x}{Consolidated binary matrix}
}
\value{
JE (Jaccard Error), EE (Euclidean Error)
}
\description{
Calculates the Jaccard and Euclidean error rates.
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
cons = consolidate(mat)
errors(cons)

}
