% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_names.R
\name{group.names}
\alias{group.names}
\title{Outputs group names specified in the input file for the creation of an nMDS plot.}
\usage{
group.names(x)
}
\arguments{
\item{x}{Consolidated binary matrix with grouping information in column 2.}
}
\value{
Scree plot.
}
\description{
Returns group names in the uploaded consolidated binary data. This will help in knowing which colours are assigned to which group name.
}
\examples{
mat = BinMatInput_ordination
group.names(mat)

}
