% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_names.R
\name{group_names}
\alias{group_names}
\title{Outputs group names specified in the input file for the creation of an nMDS plot.}
\usage{
group_names(x)
}
\arguments{
\item{x}{Consolidated binary matrix with grouping information in column 2.}
}
\value{
Scree plot.
}
\description{
Returns group names in the uploaded consolidated binary data
}
\examples{
mat = BinMatInput_ordination
group_names(mat)

}
