% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakRemove.R
\name{peakRemove}
\alias{peakRemove}
\title{Removes samples with peaks equal to or less than a specified threshold value.}
\usage{
peakRemove(x, thresh)
}
\arguments{
\item{x}{Consolidated binary matrix used for the creation of nMDS}

\item{thresh}{Peak threshold value for removal}
}
\value{
Filtered dataset
}
\description{
Removes samples with a peak number less than a specified value.
}
\examples{
 mat = BinMatInput_ordination
new = peakRemove(mat, 4)

}
