% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_replicates.R
\name{peaks.original}
\alias{peaks.original}
\title{Calculates peak numbers for the data set with all replicates (total, maximum, and minimum).}
\usage{
peaks.original(x)
}
\arguments{
\item{x}{Binary matrix comprising replicate pairs.}
}
\value{
Peak information.
}
\description{
Returns total, maximum, and minimum number of peaks in the binary matrix.
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
peaks.original(mat)

}
