% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peaks}
\alias{peaks}
\title{Calculates peak numbers (total, maximum, and minimum).}
\usage{
peaks(x)
}
\arguments{
\item{x}{Binary matrix comprising replicate pairs.}
}
\value{
Peak information.
}
\description{
Returns total, maximum, and minimum number of peaks.
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
peaks(mat)
}
