% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\name{shepard}
\alias{shepard}
\title{Creates a shepard plot.}
\usage{
shepard(x, k_val = 2, dist_meth = "binary")
}
\arguments{
\item{x}{Consolidated binary matrix.}

\item{k_val}{Number of dimensions. Set to 2 by default.}

\item{dist_meth}{Distance method. Set to "binary" by default. Other options are "euclidean", "maximum", "manhattan", "canberra", or "minkowski".}
}
\value{
Shepard plot.
}
\description{
Creates a Shepard plot for the nMDS. This indicates the 'goodness of fit' of the original distance matrix vs the ordination representation. A high R-squared value is favourable.
}
\examples{
mat = BinMatInput_ordination
shepard(mat)
}
