% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgma.R
\name{upgma}
\alias{upgma}
\title{Draws a hierarchical clustering tree (UPGMA).}
\usage{
upgma(
  x,
  bts = 10,
  size = 0.55,
  lab_size = 0.55,
  method = "binary",
  hclust = "average"
)
}
\arguments{
\item{x}{Consolidated binarx matrix}

\item{bts}{Bootstrap replications}

\item{size}{Size of plot.}

\item{lab_size}{Size of label text.}

\item{method}{Distance method. Set to 'binary' (=Jaccard distance) by default}

\item{hclust}{Clustering method. Set to 'average' (=UPGMA) by default}
}
\value{
UPGMA tree
}
\description{
Creates a UPGMA hierarchical clustering tree
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
cons = consolidate(mat)
clust = upgma(cons)

}
