% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatInput_ordination.R
\docType{data}
\name{BinMatInput_ordination}
\alias{BinMatInput_ordination}
\title{Example input data containing a consolidated binary matrix with groups}
\format{A dataframe with columns for loci, and rows of replicate pairs. Grouping information is in the second column.}
\usage{
data(BinMatInput_ordination)
}
\description{
Example input data containing a consolidated binary matrix with groups
}
\examples{
data(BinMatInput_ordination)
mat = BinMatInput_ordination
group.names(mat)
scree(mat)
shepard(mat)
clrs = c("red", "green", "black")
shp = c(16,16,16)
nmds(mat, colours = clrs, shapes = shp, labs = TRUE)
}
\keyword{datasets}
