% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate replicate pairs in a binary matrix}
\usage{
consolidate(x)
}
\arguments{
\item{x}{A CSV file containing replicate pairs of binary data.}
}
\value{
Consolidated binary matrix. For each replicate pair at each locus, 1 & 1 -> 1 (shared presence), 0 & 0 -> 0 (shared absence), 0 & 1 -> ? (ambiguity).
}
\description{
Reads in a binary matrix comprising replicate pairs.
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
cons = consolidate(mat)

}
