% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{nmds}
\alias{nmds}
\title{Creates a non-metric multidimensional scaling plot (nMDS).}
\usage{
nmds(
  x,
  dist_meth = "binary",
  k_val = 2,
  pt_size = 1,
  colours,
  shapes,
  labs = F
)
}
\arguments{
\item{x}{Consolidated binary matrix with grouping information in the second column.}

\item{dist_meth}{Distance method. Set to "binary" by default. Other options are "euclidean", "maximum", "manhattan", "canberra", or "minkowski".}

\item{k_val}{Number of dimensions for the nMDS plot.}

\item{pt_size}{Point size for symbols on the plot.}

\item{colours}{Vector containing colours to be assigned to groups.}

\item{shapes}{Vector containing pch values for shapes to be used for points.}

\item{labs}{Indicate whether labels should appear on the graph or not (T or F). Default = F.}
}
\value{
nMDS plot.
}
\description{
Creates an nMDS plot from a consolidated binary matrix with grouping information.
}
\examples{
 mat = BinMatInput_ordination
group_names(mat)
clrs = c("red", "green", "black")
shp = c(16,16,16)
nmds(mat, colours = clrs, shapes = shp, labs = TRUE)

}
