% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_consolidated.R
\name{peaks.consolidated}
\alias{peaks.consolidated}
\title{Calculates peak numbers for a consolidated data set (total, maximum, and minimum).}
\usage{
peaks.consolidated(x)
}
\arguments{
\item{x}{Binary matrix comprising replicate pairs.}
}
\value{
Peak information.
}
\description{
Returns total, maximum, and minimum number of peaks in the binary matrix.
}
\examples{
data(BinMatInput_reps)
mat = BinMatInput_reps
cons = consolidate(mat)
peaks.consolidated(cons)
}
